Moves 3DS rendering logic into sdl_video_main
