+++ src/video/sdl_video_main.c
@@ -24,6 +24,7 @@
 #include "../util.h"
 
 #ifdef __3DS__
+#include <3ds.h>
 #include <SDL3/SDL.h>
 #include <SDL3/SDL_thread.h>
 #else
@@ -38,6 +39,9 @@
 
 static void* ffmpeg_buffer;
 static size_t ffmpeg_buffer_size;
+static int surface_width, surface_height, pixel_size;
+static u8* img_buffer;
+static LightLock render_mutex;
 
 static int sdl_setup(int videoFormat, int width, int height, int redrawRate, void* context, int drFlags) {
   if (ffmpeg_init(videoFormat, width, height, SLICE_THREADING, SDL_BUFFER_FRAMES, SLICES_PER_FRAME) < 0) {
@@ -47,11 +51,116 @@ static int sdl_setup(int videoFormat, int width, int height, int redrawRate, voi
 
   ensure_buf_size(&ffmpeg_buffer, &ffmpeg_buffer_size, INITIAL_DECODER_BUFFER_SIZE + AV_INPUT_BUFFER_PADDING_SIZE);
 
+  LightLock_Init(&render_mutex);
+
+  if(y2rInit())
+  {
+    printf("Failed to initialize Y2R\n");
+    return -1;
+  }
+  Y2RU_ConversionParams y2r_parameters;
+	y2r_parameters.input_format = INPUT_YUV420_INDIV_8;
+	y2r_parameters.output_format = OUTPUT_RGB_16_565;
+	y2r_parameters.rotation = ROTATION_NONE;
+	y2r_parameters.block_alignment = BLOCK_LINE;
+	y2r_parameters.input_line_width = width;
+	y2r_parameters.input_lines = height;
+	y2r_parameters.standard_coefficient = COEFFICIENT_ITU_R_BT_709_SCALING;
+	y2r_parameters.alpha = 0xFF;
+	int status = Y2RU_SetConversionParams(&y2r_parameters);
+  if (status) {
+    printf("Failed to set Y2RU params\n");
+    return -1;
+  }
+
+  GSPGPU_FramebufferFormat px_fmt = GSP_RGB565_OES;
+  gfxInit(px_fmt, GSP_RGB565_OES, false);
+  surface_width = width;
+  surface_height = height;
+  pixel_size = gspGetBytesPerPixel(px_fmt);
+
+  img_buffer = linearAlloc(width * height * pixel_size);
+  if (!img_buffer) {
+    printf("Out of memory!");
+    return -1;
+  }
+
   return 0;
 }
 
 static void sdl_cleanup() {
   ffmpeg_destroy();
+  y2rExit();
+  linearFree(img_buffer);
+}
+
+static inline int get_dest_offset(int x, int y, int height)
+{
+  return height - y - 1 + height * x;
+}
+
+static inline int get_source_offset(int x, int y, int width)
+{
+  return x + y * width;
+}
+
+static inline void write_rgb565_to_framebuffer(u16* dest, u16* source, int width, int height) {
+  for (int y = 0; y < height; ++y) {
+    for (int x = 0; x < width; ++x) {
+      int src_offset = get_source_offset(x, y, width);
+      int dst_offset = get_dest_offset(x, y, height);
+      dest[dst_offset] = source[src_offset];
+    }
+  }
+}
+
+static inline int write_yuv_to_framebuffer(u8 *dest, const u8 **source, int width, int height) {
+	Handle conversion_finish_event_handle;
+  int status = 0;
+
+  status = Y2RU_SetSendingY(source[0], width * height, width, 0);
+  if (status) {
+    printf("Y2RU_SetSendingY failed\n");
+    goto y2ru_failed;
+  }
+
+  status = Y2RU_SetSendingU(source[1], width * height / 4, width / 2, 0);
+  if (status) {
+    printf("Y2RU_SetSendingU failed\n");
+    goto y2ru_failed;
+  }
+
+  status = Y2RU_SetSendingV(source[2], width * height / 4, width / 2, 0);
+  if (status) {
+    printf("Y2RU_SetSendingV failed\n");
+    goto y2ru_failed;
+  }
+
+  status = Y2RU_SetReceiving(img_buffer, width * height * pixel_size, width * pixel_size * 4, 0);
+  if (status) {
+    printf("Y2RU_SetReceiving failed\n");
+    goto y2ru_failed;
+  }
+
+  status = Y2RU_StartConversion();
+  if (status) {
+    printf("Y2RU_StartConversion failed\n");
+    goto y2ru_failed;
+  }
+
+  status = Y2RU_GetTransferEndEvent(&conversion_finish_event_handle);
+  if (status) {
+    printf("Y2RU_GetTransferEndEvent failed\n");
+    goto y2ru_failed;
+  }
+
+  svcWaitSynchronization(conversion_finish_event_handle, 200000000);//Wait up to 200ms.
+  svcCloseHandle(conversion_finish_event_handle);
+  write_rgb565_to_framebuffer(dest, img_buffer, width, height);
+  return DR_OK;
+
+	y2ru_failed:
+  return -1;
 }
 
 static int sdl_submit_decode_unit(PDECODE_UNIT decodeUnit) {
@@ -67,26 +176,17 @@ static int sdl_submit_decode_unit(PDECODE_UNIT decodeUnit) {
   }
   ffmpeg_decode(ffmpeg_buffer, length);
 
-  SDL_LockMutex(mutex);
   AVFrame* frame = ffmpeg_get_frame(false);
-  if (frame != NULL) {
-    sdlNextFrame++;
+  u8 *gfxtopadr = gfxGetFramebuffer(GFX_TOP, GFX_LEFT, NULL, NULL);
+  int status = write_yuv_to_framebuffer(gfxtopadr, frame->data, surface_width, surface_height);
+  gfxScreenSwapBuffers(GFX_TOP, false);
 
-    SDL_Event event;
-    event.type = SDL_EVENT_USER;
-    event.user.code = SDL_CODE_FRAME;
-    event.user.data1 = &frame->data;
-    event.user.data2 = &frame->linesize;
-    SDL_PushEvent(&event);
-  }
-  SDL_UnlockMutex(mutex);
-
-  return DR_OK;
+  return status;
 }
 
 DECODER_RENDERER_CALLBACKS decoder_callbacks_sdl = {
   .setup = sdl_setup,
   .cleanup = sdl_cleanup,
   .submitDecodeUnit = sdl_submit_decode_unit,
-  .capabilities = CAPABILITY_SLICES_PER_FRAME(SLICES_PER_FRAME) | CAPABILITY_REFERENCE_FRAME_INVALIDATION_HEVC | CAPABILITY_DIRECT_SUBMIT,
+  .capabilities = CAPABILITY_SLICES_PER_FRAME(SLICES_PER_FRAME) | CAPABILITY_REFERENCE_FRAME_INVALIDATION_HEVC,
 };
